remove registerWorkflow from DBOS to DBOSIntegration
